Improved snapshot caching
